+++ core/src/main/java/com/trendyol/design/core/preview/PreviewTheme.kt
@@ -1,11 +1,14 @@
 package com.trendyol.design.core.preview
 
+import android.content.Context
 import androidx.compose.material.Colors
 import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.ReadOnlyComposable
+import androidx.compose.runtime.remember
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
 import com.trendyol.theme.LocalTrendyolDesignTheme
 import com.trendyol.theme.TrendyolDesignColors
 import com.trendyol.theme.TrendyolDesignFontFamily
@@ -16,6 +19,12 @@ import com.trendyol.theme.TrendyolDesignTypography
 internal fun PreviewTheme(
     content: @Composable () -> Unit
 ) {
+    val context = LocalContext.current
+
+    val trendyolDesignTheme = remember {
+        createTrendyolDesignTheme(context)
+    }
+
     MaterialTheme(
         colors = TrendyolColor.materialColor
     ) {
@@ -39,8 +48,12 @@ private object TrendyolColor {
 
 private val colorPrimary = Color(color = 0xFFF27A1A)
 
-private val trendyolDesignTheme = TrendyolDesignTheme(
+private fun createTrendyolDesignTheme(context: Context) = TrendyolDesignTheme(
     colors = TrendyolDesignColors(),
     typography = TrendyolDesignTypography(),
-    fontFamily = TrendyolDesignFontFamily(),
+    fontFamily = TrendyolDesignFontFamily(
+        regular = PreviewFonts.getRobotoRegular(context),
+        medium = PreviewFonts.getRobotoMedium(context),
+        bold = PreviewFonts.getRobotoBold(context),
+    )
 )